backend/cloudrun/requirements.txt
boto3==1.34.24
fastapi==0.109.0
ffmpeg-python==0.2.0
google-auth-oauthlib==1.2.0
imageio==2.33.1
numpy==1.26.3
openai==1.9.0
opencv-python==4.9.0.80
pydantic==2.5.3
python-dotenv==1.0.0
remotion-lambda==4.0.96
requests==2.31.0
scenedetect==0.6.2
scipy==1.12.0
tqdm==4.66.1
uvicorn==0.27.0
yt-dlp==2023.12.30
